Move google maps static image to visualizations controller
