+++ lib/carto/google_maps_api.rb
@@ -5,7 +5,7 @@ module Carto
   class GoogleMapsApi
     STATIC_IMAGE_BASE_URL = 'https://maps.googleapis.com/maps/api/staticmap'.freeze
 
-    def sign(user, url)
+    def sign_url(user, url)
       raise 'User does not have Google configured' unless user.google_maps_query_string.present?
       if user.google_maps_client_id.present? && user.google_maps_private_key.present?
         # Add client=xxx + signature
@@ -43,8 +43,7 @@ module Carto
 
     def parse_basemap_styles(style_json)
       return '' unless style_json
-      styles_definition = JSON.parse(style_json, symbolize_names: true)
-      styles = styles_definition.map do |style_definition|
+      styles = style_json.map do |style_definition|
         style_parts = []
         style_parts << "feature:#{style_definition[:featureType] || 'all'}"
         style_parts << "element:#{style_definition[:elementType] || 'all'}"
