+++ spec/requests/carto/api/visualizations_controller_spec.rb
@@ -2503,6 +2503,51 @@ describe Carto::Api::VisualizationsController do
     end
   end
 
+  describe '#google_maps_static_image' do
+    before(:all) do
+      @user = FactoryGirl.create(:carto_user)
+      @visualization = FactoryGirl.create(:carto_visualization, user: @user)
+    end
+
+    before(:each) do
+      host! "#{@user.username}.localhost.lan"
+      login_as(@user, scope: @user.username)
+    end
+
+    after(:all) do
+      @visualization.destroy
+      @user.destroy
+    end
+
+    let(:params) do
+      {
+        mapType: 'roadmap',
+        size: '300x200',
+        zoom: 14,
+        center: '0.12,-7.56',
+        style: []
+      }
+    end
+
+    it 'returns error if user does not have Google configured' do
+      @user.google_maps_key = nil
+      @user.save
+      post_json api_v1_google_maps_static_image_url(id: @visualization.id), params do |response|
+        expect(response.status).to eq 400
+        expect(response.body[:errors]).to be
+      end
+    end
+
+    it 'returns signed google maps URL' do
+      @user.google_maps_key = 'key=GAdhfasjkd'
+      @user.save
+      post_json api_v1_google_maps_static_image_url(id: @visualization.id), params do |response|
+        response.status.should be_success
+        response.body[:url].should eq 'https://maps.googleapis.com/maps/api/staticmap?center=0.12,-7.56&mapType=roadmap&size=300x200&zoom=14&key=GAdhfasjkd'
+      end
+    end
+  end
+
   include Rack::Test::Methods
   include Warden::Test::Helpers
   include CacheHelper
